add memcpy

+++ ft_strlen.c
@@ -12,12 +12,12 @@
 
 #include "libft.h"
 
-int	ft_strlen(const char *s)
+size_t	ft_strlen(const char *s)
 {
-	int	i;
+	unsigned int	i;
 	
 	i = 0;
 	while (s[i] != '\0')
 		i++;
-	return (i);
+	return ((size_t)i);
 }
